Move DefaultJavaScriptEncoder to separate file and override EncodeUtf8 with better caching.
